Move computeTimeStampsWithTorqueLimits to new file
